# docs/openapi/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台管理接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/subscription/stripe/pay": {
      "post": {
        "summary": "创建版本化 Stripe 套餐单次购买",
        "tags": [
          "订阅套餐"
        ],
        "description": "新模式必须提供 version_id 和 event_id。身份取认证上下文，订单先保存并锁价。USD/CNY 单次付款、创建发票并使用原订单号作渠道防重标识；不自动续扣。回调服务端查询实际发票付款时间后开通，缺事实保留核查。",
        "security": [
          {
            "AccessToken1": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {}
                  }
                }
              }
            }
          },
          "400": {
            "description": "输入不合法"
          },
          "401": {
            "description": "身份无效、令牌过期或已撤销"
          },
          "403": {
            "description": "角色、用途权限或资源权限不足"
          },
          "404": {
            "description": "记录不存在"
          },
          "409": {
            "description": "重复事件输入冲突、订阅不可操作或可用积分不足"
          },
          "500": {
            "description": "存储失败；不暴露内部错误"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "anyOf": [
                  {
                    "required": [
                      "version_id",
                      "event_id"
                    ]
                  },
                  {
                    "required": [
                      "plan_id"
                    ]
                  }
                ],
                "properties": {
                  "version_id": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "event_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 128
                  },
                  "plan_id": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "operationId": "SubscriptionRequestStripePay"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "SessionAuth1": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "已废弃：Gin session 不再用于面板鉴权"
      },
      "AccessToken1": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or PAT",
        "description": "面板短期 JWT 或 User.AccessToken PAT；不再要求 New-Api-User"
      },
      "NewApiUser1": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "已废弃且不再参与鉴权；PAT 客户端可以移除此请求头"
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "new_api_refresh",
        "description": "仅 refresh/logout 使用的 HttpOnly、SameSite=Strict Cookie"
      },
      "SecurityProof": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Security-Proof",
        "description": "绑定用户、登录会话和 scope 的短期二次验证 Proof"
      },
      "Combination222": {
        "group": [
          {
            "id": 573667
          }
        ],
        "type": "combination"
      },
      "Combination1122": {
        "group": [
...
  },
  "servers": [],
  "security": [
    {
      "Combination343": []
    },
    {
      "Combination1243": []
    }
  ]
}